feat: add separate benchmark tasks
