+++ servers/graphql-kotlin-spring-server/src/test/kotlin/com/expediagroup/graphql/server/spring/SubscriptionConfigurationTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Expedia, Inc
+ * Copyright 2024 Expedia, Inc
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,6 +29,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import graphql.GraphQL
 import graphql.schema.GraphQLSchema
+import io.mockk.every
 import io.mockk.mockk
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -36,7 +37,6 @@ import org.springframework.boot.autoconfigure.AutoConfigurations
 import org.springframework.boot.test.context.runner.ReactiveWebApplicationContextRunner
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.HandlerMapping
 import org.springframework.web.reactive.socket.server.support.WebSocketHandlerAdapter
 import reactor.core.publisher.Flux
 import java.time.Duration
@@ -73,7 +73,7 @@ class SubscriptionConfigurationTest {
 
                 assertThat(ctx).hasSingleBean(SubscriptionWebSocketHandler::class.java)
                 assertThat(ctx).hasSingleBean(WebSocketHandlerAdapter::class.java)
-                assertThat(ctx).hasSingleBean(HandlerMapping::class.java)
+                assertThat(ctx).hasBean("subscriptionHandlerMapping")
             }
     }
 
@@ -102,10 +102,9 @@ class SubscriptionConfigurationTest {
 
                 assertThat(ctx).hasSingleBean(ApolloSubscriptionWebSocketHandler::class.java)
 
-                assertThat(ctx).hasSingleBean(WebSocketHandlerAdapter::class.java)
-                assertThat(ctx).getBean(WebSocketHandlerAdapter::class.java)
+                assertThat(ctx).hasBean("webSocketHandlerAdapter")
+                assertThat(ctx).getBean("webSocketHandlerAdapter")
                     .isSameAs(customConfiguration.webSocketHandlerAdapter())
-                assertThat(ctx).hasSingleBean(HandlerMapping::class.java)
             }
     }
 
@@ -146,7 +145,9 @@ class SubscriptionConfigurationTest {
         @Bean
         fun subscription(): Subscription = SimpleSubscription()
         @Bean
-        fun webSocketHandlerAdapter(): WebSocketHandlerAdapter = mockk()
+        fun webSocketHandlerAdapter(): WebSocketHandlerAdapter = mockk {
+            every { order } returns 1
+        }
     }
 
     // GraphQL spec requires at least single query to be present as Query type is needed to run introspection queries
